refactor: put Alarm in Notice folder and remove LogoutButton
