+++ front/src/components/NavigationBar/DefaultNavBtnContainer/DefaultNavBtnContainer.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import APP_URI from 'util/uri';
 import { faUserAlt, faSignOutAlt } from '@fortawesome/free-solid-svg-icons';
-import Notice from 'components/NavigationBar/DefaultNavBtnContainer/Notice';
+import Notice from 'components/NavigationBar/DefaultNavBtnContainer/Notice/Notice';
 import { NavBtnContainerStyle, NavImageStyle } from 'components/NavigationBar/DefaultNavBtnContainer/DefaultNavBtnContainer.style';
 
 export default function DefaultNavBtnContainer() {
