+++ front/src/components/NavigationBar/DefaultNavBtnContainer/Notice/Notice.js
@@ -1,5 +1,5 @@
 import React, { useReducer, useEffect } from 'react';
-import Alarm from 'components/Alarm/Alarm';
+import Alarm from 'components/NavigationBar/DefaultNavBtnContainer/Notice/Alarm';
 import Messages from 'components/Messages/Messages';
 import { faBell } from '@fortawesome/free-solid-svg-icons';
 import { NavImageStyle } from 'components/NavigationBar/DefaultNavBtnContainer/DefaultNavBtnContainer.style';
